Use sign of the electric charge of gen particles
The issue is that TParticlePDG::Charge() returns its electric charge *
3, so that the charge of quarks is an integer. It also means that gen
electrons have charge +/-3, which causes the charge-based gen-matching
to fail when applying lepton ID SFs, thus changing the data/MC SF and
the eventual event weights. The solution is to assign just the sign of
the gen particle as its charge, since we don't care about the fractional
nature of quark charges.