Read hadron (and parton) flavor of reconstructed jets
Needed for assigning b-tagging SF. Closes perfectly with the
nanoaod-tools implementation, except for the b-tagging SF that are
computed by shifting the JES scale, which causes subtle bin-by-bin
migrations. In fact, the implementation here is more correct, because I
had a bug in the nanoaod-tools implementation, where only the nominal
jet pT was used as input to the b-tagging SF evaluator.
Also, not reading jet charges anymore. Removed all references to
b-taging SF ratios, because these corrections should be derived and
applied at the analysis level.