Fix spurious MET
Some JES uncertainties returned by correctionlib had value of 999,
likely because the input jet was very forward, at abs(eta) > 5.4, even
though the JSON API claims that overflow is ok. Need to follow up on
this with the correctionlib author and/or JetMET group to better
understand how to deal with potential overflows of the input parameters.
Also, added debugging to the JMECorrector class.