Propagate JEC to AK4 jets and MET correctly
Previously, the JEC was computed from the same raw jet pT at every
level, but the correct thing is to use the JEC obtained from the
previous level and apply it to the inputs at the current level.
Additionally, the compound JEC as well as L1 JEC are applicable only to
the raw jet. This was not correctly done previously when propagating the
effects of JEC to MET. Implementing the said changes improved closure
with the NanoAOD as well as nanoAOD-tools to a perfect level. The
testing was based on 2018 HH signal MC, which has now been made the
default.